improved import of update_symplectic
