+++ strawberryfields/backends/tfbackend/__init__.py
@@ -150,22 +150,6 @@ else:
     tf_available = True
     tf_version = tensorflow.__version__
 
-if tf_available:
-
-    def update_symplectic(S, dS, lr):
-        """returns the updated syplectic matrix S according to its geodesic.
-        S (Tensor): symplectic matrix to be updated.
-        dS (Tensor): euclidean gradient of S.
-        lr (float): learning rate.
-        """
-        Jmat = sympmat(S.shape[1] // 2)
-        Z = np.matmul(np.transpose(S), dS)
-        Y = 0.5 * (Z + np.linalg.multi_dot([Jmat, Z.T, Jmat]))
-        S.assign(
-            S @ expm(-lr * np.transpose(Y)) @ expm(-lr * (Y - np.transpose(Y))), read_value=False
-        )
-
-
 tf_info = """\
 To use Strawberry Fields with TensorFlow support, version 2.x of TensorFlow is required.
 This can be installed as follows:
@@ -184,3 +168,6 @@ if not (tf_available and tf_version[:2] == "2."):
     sys.excepthook = excepthook
 
     raise ImportError(tf_info)
+
+
+from .ops import update_symplectic
